Attempt to get all segment users in API call at once
This change is fixing issue where expensive segments (couple of seconds
to return) with many users were bombing the DB requiring the expensive
query to be executed first and then returning only 1000 users from the
list instead of everybody.

The idea here is that since the user list was already calculated, it
doesn't really make sense to throw it away and let ->process() to
iterate over 1000 users again.

As memory might have been the issue in the past to introduce 1000 user
steps, we've added memory information to the API call response.